feat: abstract types
